добавлены интерфейсы для сервисов в модуле auth-service
