+++ Bazzar/auth-service/src/main/java/ru/bazzar/auth/services/RoleService.java
@@ -1,19 +1,7 @@
 package ru.bazzar.auth.services;
 
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Component;
-import ru.bazzar.api.ResourceNotFoundException;
 import ru.bazzar.auth.entities.Role;
-import ru.bazzar.auth.repositories.RoleRepository;
 
-@Component
-@RequiredArgsConstructor
-public class RoleService {
-
-    private final RoleRepository roleRepository;
-
-    public Role findRoleByName(String name) {
-        return roleRepository.findRoleByName(name).orElseThrow(() -> new ResourceNotFoundException("Роль с id: " + name + " не найдена!"));
-    }
-
+public interface RoleService {
+    Role findRoleByName(String name);
 }
