+++ Bazzar/auth-service/src/main/java/ru/bazzar/auth/services/RoleServiceImpl.java
@@ -1,0 +1,19 @@
+package ru.bazzar.auth.services;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Component;
+import ru.bazzar.api.ResourceNotFoundException;
+import ru.bazzar.auth.entities.Role;
+import ru.bazzar.auth.repositories.RoleRepository;
+
+@Component
+@RequiredArgsConstructor
+public class RoleServiceImpl implements RoleService {
+
+    private final RoleRepository roleRepository;
+
+    public Role findRoleByName(String name) {
+        return roleRepository.findRoleByName(name).orElseThrow(() -> new ResourceNotFoundException("Роль с id: " + name + " не найдена!"));
+    }
+
+}
